projects/advanced_rest_features/requirements.txt
Flask==3.0.0
google-cloud-datastore==2.15.1
json2html==1.3.0 